Update sample apps for encoding interface config
Includes two boilerplate apps and seven custom apps to encode interface
configuration in XML:
cd-encode-xr-ifmgr-cfg-10-ydk.py - boilerplate (global)
cd-encode-xr-ifmgr-cfg-11-ydk.py - boilerplate
cd-encode-xr-ifmgr-cfg-20-ydk.py - default message behavior
cd-encode-xr-ifmgr-cfg-22-ydk.py - disable link messages
cd-encode-xr-ifmgr-cfg-24-ydk.py - link messages for logical interfaces
cd-encode-xr-ifmgr-cfg-30-ydk.py - virtual IPv4 intf (lo0)
cd-encode-xr-ifmgr-cfg-32-ydk.py - virtual IPv6 intf (lo0)
cd-encode-xr-ifmgr-cfg-34-ydk.py - physical IPv4 intf (ge0/0/0/0)
cd-encode-xr-ifmgr-cfg-36-ydk.py - physical IPv6 intf (ge0/0/0/0)
Refactors previous sample apps to make them compatible with model
bundles introduced in YDK 0.5.0.  App names now use a notation that is
more service and model agnostic.  Previous apps were committed in
https://github.com/CiscoDevNet/ydk-py-samples/commit/861c1f6a22792df599ed47fcc6a639a052f18529
https://github.com/CiscoDevNet/ydk-py-samples/commit/22746a9b349b3b57ee7e8b63735a5c9e2356df96
