+++ samples/basic/codec/models/cisco-ios-xr/Cisco-IOS-XR-ifmgr-cfg/cd-encode-xr-ifmgr-cfg-10-ydk.py
@@ -16,9 +16,9 @@
 #
 
 """
-Encode config for model Cisco-IOS-XR-ifmgr-cfg.
+Encode configuration for model Cisco-IOS-XR-ifmgr-cfg.
 
-usage: cd-encode-config-ifmgr-20-ydk.py [-h] [-v]
+usage: cd-encode-xr-ifmgr-cfg-10-ydk.py [-h] [-v]
 
 optional arguments:
   -h, --help     show this help message and exit
@@ -30,14 +30,14 @@ from urlparse import urlparse
 
 from ydk.services import CodecService
 from ydk.providers import CodecServiceProvider
-from ydk.models.ifmgr import Cisco_IOS_XR_ifmgr_cfg as xr_ifmgr_cfg
+from ydk.models.cisco_ios_xr import Cisco_IOS_XR_ifmgr_cfg \
+    as xr_ifmgr_cfg
 import logging
 
 
 def config_global_interface_configuration(global_interface_configuration):
     """Add config data to global_interface_configuration object."""
-    # display link status messages for physical links
-    global_interface_configuration.link_status = xr_ifmgr_cfg.LinkStatusEnumEnum.DEFAULT
+    pass
 
 
 if __name__ == "__main__":
@@ -63,13 +63,12 @@ if __name__ == "__main__":
     # create codec service
     codec = CodecService()
 
-    # create config object
-    global_interface_configuration = xr_ifmgr_cfg.GlobalInterfaceConfiguration()
-    # add object configuration
-    config_global_interface_configuration(global_interface_configuration)
+    global_interface_configuration = xr_ifmgr_cfg.GlobalInterfaceConfiguration()  # create object
+    config_global_interface_configuration(global_interface_configuration)  # add object configuration
 
     # encode and print object
-    print(codec.encode(provider, global_interface_configuration))
+    # print(codec.encode(provider, global_interface_configuration))
+
     provider.close()
     exit()
 # End of script
