+++ samples/basic/codec/models/cisco-ios-xr/Cisco-IOS-XR-ifmgr-cfg/cd-encode-xr-ifmgr-cfg-20-ydk.py
@@ -16,9 +16,9 @@
 #
 
 """
-Encode config for model Cisco-IOS-XR-ifmgr-cfg.
+Encode configuration for model Cisco-IOS-XR-ifmgr-cfg.
 
-usage: cd-encode-config-ifmgr-36-ydk.py [-h] [-v]
+usage: cd-encode-xr-ifmgr-cfg-20-ydk.py [-h] [-v]
 
 optional arguments:
   -h, --help     show this help message and exit
@@ -30,28 +30,15 @@ from urlparse import urlparse
 
 from ydk.services import CodecService
 from ydk.providers import CodecServiceProvider
-from ydk.models.ifmgr import Cisco_IOS_XR_ifmgr_cfg as xr_ifmgr_cfg
+from ydk.models.cisco_ios_xr import Cisco_IOS_XR_ifmgr_cfg \
+    as xr_ifmgr_cfg
 import logging
 
 
-def config_interface_configurations(interface_configurations):
-    """Add config data to interface_configurations object."""
-    # configure IPv6 loopback
-    interface_configuration = interface_configurations.InterfaceConfiguration()
-    interface_configuration.active = "act"
-    interface_configuration.interface_name = "GigabitEthernet0/0/0/0"
-    interface_configuration.description = "CONNECTS TO LSR1 (g0/0/0/1)"
-    mtu = interface_configuration.mtus.Mtu()
-    mtu.owner = "GigabitEthernet"
-    mtu.mtu = 9192
-    interface_configuration.mtus.mtu.append(mtu)
-    addresses = interface_configuration.ipv6_network.addresses
-    regular_address = addresses.regular_addresses.RegularAddress()
-    regular_address.address = "2001:db8::1:0"
-    regular_address.prefix_length = 127
-    addresses.regular_addresses.regular_address.append(regular_address)
-    interface_configuration.statistics.load_interval = 30
-    interface_configurations.interface_configuration.append(interface_configuration)
+def config_global_interface_configuration(global_interface_configuration):
+    """Add config data to global_interface_configuration object."""
+    # display link status messages for physical links
+    global_interface_configuration.link_status = xr_ifmgr_cfg.LinkStatusEnumEnum.DEFAULT
 
 
 if __name__ == "__main__":
@@ -77,13 +64,12 @@ if __name__ == "__main__":
     # create codec service
     codec = CodecService()
 
-    # create config object
-    interface_configurations = xr_ifmgr_cfg.InterfaceConfigurations()
-    # add object configuration
-    config_interface_configurations(interface_configurations)
+    global_interface_configuration = xr_ifmgr_cfg.GlobalInterfaceConfiguration()  # create object
+    config_global_interface_configuration(global_interface_configuration)  # add object configuration
 
     # encode and print object
-    print(codec.encode(provider, interface_configurations))
+    print(codec.encode(provider, global_interface_configuration))
+
     provider.close()
     exit()
 # End of script
