+++ samples/basic/codec/models/cisco-ios-xr/Cisco-IOS-XR-ifmgr-cfg/cd-encode-xr-ifmgr-cfg-34-ydk.xml
