+++ samples/basic/codec/models/cisco-ios-xr/Cisco-IOS-XR-ifmgr-cfg/cd-encode-xr-ifmgr-cfg-36-ydk.py
@@ -16,9 +16,9 @@
 #
 
 """
-Encode config for model Cisco-IOS-XR-ifmgr-cfg.
+Encode configuration for model Cisco-IOS-XR-ifmgr-cfg.
 
-usage: cd-encode-config-ifmgr-34-ydk.py [-h] [-v]
+usage: cd-encode-xr-ifmgr-cfg-36-ydk.py [-h] [-v]
 
 optional arguments:
   -h, --help     show this help message and exit
@@ -30,13 +30,14 @@ from urlparse import urlparse
 
 from ydk.services import CodecService
 from ydk.providers import CodecServiceProvider
-from ydk.models.ifmgr import Cisco_IOS_XR_ifmgr_cfg as xr_ifmgr_cfg
+from ydk.models.cisco_ios_xr import Cisco_IOS_XR_ifmgr_cfg \
+    as xr_ifmgr_cfg
 import logging
 
 
 def config_interface_configurations(interface_configurations):
     """Add config data to interface_configurations object."""
-    # configure IPv4 loopback
+    # configure IPv6 loopback
     interface_configuration = interface_configurations.InterfaceConfiguration()
     interface_configuration.active = "act"
     interface_configuration.interface_name = "GigabitEthernet0/0/0/0"
@@ -45,10 +46,11 @@ def config_interface_configurations(interface_configurations):
     mtu.owner = "GigabitEthernet"
     mtu.mtu = 9192
     interface_configuration.mtus.mtu.append(mtu)
-    primary = interface_configuration.ipv4_network.addresses.Primary()
-    primary.address = "172.16.1.0"
-    primary.netmask = "255.255.255.254"
-    interface_configuration.ipv4_network.addresses.primary = primary
+    addresses = interface_configuration.ipv6_network.addresses
+    regular_address = addresses.regular_addresses.RegularAddress()
+    regular_address.address = "2001:db8::1:0"
+    regular_address.prefix_length = 127
+    addresses.regular_addresses.regular_address.append(regular_address)
     interface_configuration.statistics.load_interval = 30
     interface_configurations.interface_configuration.append(interface_configuration)
 
@@ -76,13 +78,12 @@ if __name__ == "__main__":
     # create codec service
     codec = CodecService()
 
-    # create config object
-    interface_configurations = xr_ifmgr_cfg.InterfaceConfigurations()
-    # add object configuration
-    config_interface_configurations(interface_configurations)
+    interface_configurations = xr_ifmgr_cfg.InterfaceConfigurations()  # create object
+    config_interface_configurations(interface_configurations)  # add object configuration
 
     # encode and print object
     print(codec.encode(provider, interface_configurations))
+
     provider.close()
     exit()
 # End of script
